refactor:ClientSideComponentコンポーネントを移動
