+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import Footer from "../components/organisms/footer";
-import ClientSideComponent from '../components/ClientSideComponent';
+import ClientSideComponent from '../components/organisms/ClientSideComponent';
 
 const inter = Inter({ subsets: ["latin"] });
